begin to implement normals computing

+++ src/algebra3.cpp
@@ -102,7 +102,7 @@ vec2::vec2()
     n[VX] = n[VY] = 0.0;
 }
 
-vec2::vec2(float x, float y)
+vec2::vec2(double x, double y)
 {
     n[VX] = x;
     n[VY] = y;
@@ -153,22 +153,22 @@ vec2 & vec2::operator-=(const vec2 &v)
     return *this;
 }
 
-vec2 &vec2::operator*=(float d)
+vec2 &vec2::operator*=(double d)
 {
     n[VX] *= d;
     n[VY] *= d;
     return *this;
 }
 
-vec2 &vec2::operator/=(float d)
+vec2 &vec2::operator/=(double d)
 {
-    float d_inv = 1.0f/d;
+    double d_inv = 1.0f/d;
     n[VX] *= d_inv;
     n[VY] *= d_inv;
     return *this;
 }
 
-float &vec2::operator[](int i)
+double &vec2::operator[](int i)
 {
     if (i < VX || i > VY)
       //VEC_ERROR("vec2 [] operator: illegal access; index = " << i << '\n')
@@ -176,7 +176,7 @@ float &vec2::operator[](int i)
     return n[i];
 }
 
-const float &vec2::operator[](int i) const
+const double &vec2::operator[](int i) const
 {
     if (i < VX || i > VY)
       //VEC_ERROR("vec2 [] operator: illegal access; index = " << i << '\n')
@@ -187,12 +187,12 @@ const float &vec2::operator[](int i) const
 
 /******************** vec2 SPECIAL FUNCTIONS ********************/
 
-float vec2::length() const
+double vec2::length() const
 {
-    return (float) sqrt(length2());
+    return (double) sqrt(length2());
 }
 
-float vec2::length2() const
+double vec2::length2() const
 {
     return n[VX]*n[VX] + n[VY]*n[VY];
 }
@@ -210,7 +210,7 @@ vec2 &vec2::apply(V_FCT_PTR fct)
     return *this;
 }
 
-void vec2::set( float x, float y )
+void vec2::set( double x, double y )
 {
   n[VX] = x;   n[VY] = y;
 }
@@ -232,12 +232,12 @@ vec2 GLUI::operator-(const vec2 &a, const vec2& b)
     return vec2(a.n[VX]-b.n[VX], a.n[VY]-b.n[VY]);
 }
 
-vec2 GLUI::operator*(const vec2 &a, float d)
+vec2 GLUI::operator*(const vec2 &a, double d)
 {
     return vec2(d*a.n[VX], d*a.n[VY]);
 }
 
-vec2 GLUI::operator*(float d, const vec2 &a)
+vec2 GLUI::operator*(double d, const vec2 &a)
 {
     return a*d;
 }
@@ -274,14 +274,14 @@ vec3 GLUI::operator*(const vec3 &v, const mat3 &a)
     return a.transpose() * v;
 }
 
-float GLUI::operator*(const vec2 &a, const vec2 &b)
+double GLUI::operator*(const vec2 &a, const vec2 &b)
 {
     return a.n[VX]*b.n[VX] + a.n[VY]*b.n[VY];
 }
 
-vec2 GLUI::operator/(const vec2 &a, float d)
+vec2 GLUI::operator/(const vec2 &a, double d)
 {
-    float d_inv = 1.0f/d;
+    double d_inv = 1.0f/d;
     return vec2(a.n[VX]*d_inv, a.n[VY]*d_inv);
 }
 
@@ -362,14 +362,14 @@ vec3::vec3()
     n[VX] = n[VY] = n[VZ] = 0.0;
 }
 
-vec3::vec3(float array[3])
+vec3::vec3(double array[3])
 {
     n[VX] = array[0];
     n[VY] = array[1];
     n[VZ] = array[2];
 }
 
-vec3::vec3(float x, float y, float z)
+vec3::vec3(double x, double y, double z)
 {
     n[VX] = x;
     n[VY] = y;
@@ -388,7 +388,7 @@ vec3::vec3(const vec2 &v)
     n[VZ] = 1.0;
 }
 
-vec3::vec3(const vec2 &v, float d)
+vec3::vec3(const vec2 &v, double d)
 {
     n[VX] = v.n[VX];
     n[VY] = v.n[VY];
@@ -440,7 +440,7 @@ vec3 &vec3::operator-=(const vec3& v)
     return *this;
 }
 
-vec3 &vec3::operator*=(float d)
+vec3 &vec3::operator*=(double d)
 {
     n[VX] *= d;
     n[VY] *= d;
@@ -448,16 +448,16 @@ vec3 &vec3::operator*=(float d)
     return *this;
 }
 
-vec3 &vec3::operator/=(float d)
+vec3 &vec3::operator/=(double d)
 {
-    float d_inv = 1.0f/d;
+    double d_inv = 1.0f/d;
     n[VX] *= d_inv;
     n[VY] *= d_inv;
     n[VZ] *= d_inv;
     return *this;
 }
 
-float &vec3::operator[](int i)
+double &vec3::operator[](int i)
 {
     if (i < VX || i > VZ)
         //VEC_ERROR("vec3 [] operator: illegal access; index = " << i << '\n')
@@ -466,7 +466,7 @@ float &vec3::operator[](int i)
     return n[i];
 }
 
-const float &vec3::operator[](int i) const
+const double &vec3::operator[](int i) const
 {
     if (i < VX || i > VZ)
         //VEC_ERROR("vec3 [] operator: illegal access; index = " << i << '\n')
@@ -477,12 +477,12 @@ const float &vec3::operator[](int i) const
 
 // SPECIAL FUNCTIONS
 
-float vec3::length() const
+double vec3::length() const
 {
-    return (float) sqrt(length2());
+    return (double) sqrt(length2());
 }
 
-float vec3::length2() const
+double vec3::length2() const
 {
     return n[VX]*n[VX] + n[VY]*n[VY] + n[VZ]*n[VZ];
 }
@@ -509,7 +509,7 @@ vec3 &vec3::apply(V_FCT_PTR fct)
     return *this;
 }
 
-void vec3::set(float x, float y, float z)   // set vector
+void vec3::set(double x, double y, double z)   // set vector
 {
     n[VX] = x;
     n[VY] = y;
@@ -538,12 +538,12 @@ vec3 GLUI::operator-(const vec3 &a, const vec3 &b)
     return vec3(a.n[VX]-b.n[VX], a.n[VY]-b.n[VY], a.n[VZ]-b.n[VZ]);
 }
 
-vec3 GLUI::operator*(const vec3 &a, float d)
+vec3 GLUI::operator*(const vec3 &a, double d)
 {
     return vec3(d*a.n[VX], d*a.n[VY], d*a.n[VZ]);
 }
 
-vec3 GLUI::operator*(float d, const vec3 &a)
+vec3 GLUI::operator*(double d, const vec3 &a)
 {
     return a*d;
 }
@@ -558,14 +558,14 @@ vec3 GLUI::operator*(const mat4 &a, const vec3 &v)
     return a.transpose()*v;
 }*/
 
-float GLUI::operator*(const vec3 &a, const vec3 &b)
+double GLUI::operator*(const vec3 &a, const vec3 &b)
 {
     return a.n[VX]*b.n[VX] + a.n[VY]*b.n[VY] + a.n[VZ]*b.n[VZ];
 }
 
-vec3 GLUI::operator/(const vec3 &a, float d)
+vec3 GLUI::operator/(const vec3 &a, double d)
 {
-    float d_inv = 1.0f/d;
+    double d_inv = 1.0f/d;
     return vec3(a.n[VX]*d_inv, a.n[VY]*d_inv, a.n[VZ]*d_inv);
 }
 
@@ -655,7 +655,7 @@ vec4::vec4()
     n[VW] = 1.0;
 }
 
-vec4::vec4(float x, float y, float z, float w)
+vec4::vec4(double x, double y, double z, double w)
 {
     n[VX] = x;
     n[VY] = y;
@@ -679,7 +679,7 @@ vec4::vec4(const vec3 &v)
     n[VW] = 1.0;
 }
 
-vec4::vec4(const vec3 &v, float d)
+vec4::vec4(const vec3 &v, double d)
 {
     n[VX] = v.n[VX];
     n[VY] = v.n[VY];
@@ -716,7 +716,7 @@ vec4 &vec4::operator-=(const vec4 &v)
     return *this;
 }
 
-vec4 &vec4::operator*=(float d)
+vec4 &vec4::operator*=(double d)
 {
     n[VX] *= d;
     n[VY] *= d;
@@ -725,9 +725,9 @@ vec4 &vec4::operator*=(float d)
     return *this;
 }
 
-vec4 &vec4::operator/=(float d)
+vec4 &vec4::operator/=(double d)
 {
-    float d_inv = 1.0f/d;
+    double d_inv = 1.0f/d;
     n[VX] *= d_inv;
     n[VY] *= d_inv;
     n[VZ] *= d_inv;
@@ -735,7 +735,7 @@ vec4 &vec4::operator/=(float d)
     return *this;
 }
 
-float &vec4::operator[](int i)
+double &vec4::operator[](int i)
 {
     if (i < VX || i > VW)
         //VEC_ERROR("vec4 [] operator: illegal access; index = " << i << '\n')
@@ -744,7 +744,7 @@ float &vec4::operator[](int i)
     return n[i];
 }
 
-const float &vec4::operator[](int i) const
+const double &vec4::operator[](int i) const
 {
     if (i < VX || i > VW)
         //VEC_ERROR("vec4 [] operator: illegal access; index = " << i << '\n')
@@ -755,12 +755,12 @@ const float &vec4::operator[](int i) const
 
 // SPECIAL FUNCTIONS
 
-float vec4::length() const
+double vec4::length() const
 {
-    return (float) sqrt(length2());
+    return (double) sqrt(length2());
 }
 
-float vec4::length2() const
+double vec4::length2() const
 {
     return n[VX]*n[VX] + n[VY]*n[VY] + n[VZ]*n[VZ] + n[VW]*n[VW];
 }
@@ -794,7 +794,7 @@ void vec4::print(FILE *file, const char *name) const // print vector to a file
     fprintf( file, "%s: <%f, %f, %f, %f>\n", name, n[VX], n[VY], n[VZ], n[VW]);
 }
 
-void vec4::set(float x, float y, float z, float a)
+void vec4::set(double x, double y, double z, double a)
 {
     n[0] = x;
     n[1] = y;
@@ -828,12 +828,12 @@ vec4 GLUI::operator-(const vec4 &a, const vec4 &b)
         a.n[VW] - b.n[VW]);
 }
 
-vec4 GLUI::operator*(const vec4 &a, float d)
+vec4 GLUI::operator*(const vec4 &a, double d)
 {
     return vec4(d*a.n[VX], d*a.n[VY], d*a.n[VZ], d*a.n[VW]);
 }
 
-vec4 GLUI::operator*(float d, const vec4 &a)
+vec4 GLUI::operator*(double d, const vec4 &a)
 {
     return a*d;
 }
@@ -856,7 +856,7 @@ vec4 GLUI::operator*(const vec4 &v, const mat4 &a)
     return a.transpose()*v;
 }
 
-float GLUI::operator*(const vec4 &a, const vec4 &b)
+double GLUI::operator*(const vec4 &a, const vec4 &b)
 {
     return
         a.n[VX]*b.n[VX] +
@@ -865,9 +865,9 @@ float GLUI::operator*(const vec4 &a, const vec4 &b)
         a.n[VW]*b.n[VW];
 }
 
-vec4 GLUI::operator/(const vec4 &a, float d)
+vec4 GLUI::operator/(const vec4 &a, double d)
 {
-    float d_inv = 1.0f/d;
+    double d_inv = 1.0f/d;
     return vec4(
         a.n[VX]*d_inv,
         a.n[VY]*d_inv,
@@ -1001,7 +1001,7 @@ mat3 &mat3::operator-=(const mat3& m)
     return *this;
 }
 
-mat3 &mat3::operator*=(float d)
+mat3 &mat3::operator*=(double d)
 {
     v[0] *= d;
     v[1] *= d;
@@ -1009,7 +1009,7 @@ mat3 &mat3::operator*=(float d)
     return *this;
 }
 
-mat3 &mat3::operator/=(float d)
+mat3 &mat3::operator/=(double d)
 {
     v[0] /= d;
     v[1] /= d;
@@ -1128,17 +1128,17 @@ mat3 GLUI::operator*(const mat3 &a, const mat3 &b)
     #undef ROWCOL
 }
 
-mat3 GLUI::operator*(const mat3 &a, float d)
+mat3 GLUI::operator*(const mat3 &a, double d)
 {
     return mat3(a.v[0]*d, a.v[1]*d, a.v[2]*d);
 }
 
-mat3 GLUI::operator*(float d, const mat3 &a)
+mat3 GLUI::operator*(double d, const mat3 &a)
 {
     return a*d;
 }
 
-mat3 GLUI::operator/(const mat3 &a, float d)
+mat3 GLUI::operator/(const mat3 &a, double d)
 {
     return mat3(a.v[0]/d, a.v[1]/d, a.v[2]/d);
 }
@@ -1225,10 +1225,10 @@ mat4::mat4(const mat4 &m)
 }
 
 mat4::mat4(
-     float a00, float a01, float a02, float a03,
-     float a10, float a11, float a12, float a13,
-     float a20, float a21, float a22, float a23,
-     float a30, float a31, float a32, float a33 )
+     double a00, double a01, double a02, double a03,
+     double a10, double a11, double a12, double a13,
+     double a20, double a21, double a22, double a23,
+     double a30, double a31, double a32, double a33 )
 {
   v[0][0] = a00;  v[0][1] = a01;  v[0][2] = a02;  v[0][3] = a03;
   v[1][0] = a10;  v[1][1] = a11;  v[1][2] = a12;  v[1][3] = a13;
@@ -1265,7 +1265,7 @@ mat4 &mat4::operator-=(const mat4 &m)
     return *this;
 }
 
-mat4 &mat4::operator*=(float d)
+mat4 &mat4::operator*=(double d)
 {
     v[0] *= d;
     v[1] *= d;
@@ -1274,7 +1274,7 @@ mat4 &mat4::operator*=(float d)
     return *this;
 }
 
-mat4 &mat4::operator/=(float d)
+mat4 &mat4::operator/=(double d)
 {
     v[0] /= d;
     v[1] /= d;
@@ -1384,7 +1384,7 @@ void mat4::swap_rows(int i, int j)
 
 void mat4::swap_cols(int i, int j)
 {
-    float t;
+    double t;
     int k;
 
     for (k=0; k<4; k++)
@@ -1439,17 +1439,17 @@ mat4 GLUI::operator*(const mat4 &a, const mat4 &b)
     #undef ROWCOL
 }
 
-mat4 GLUI::operator*(const mat4 &a, float d)
+mat4 GLUI::operator*(const mat4 &a, double d)
 {
     return mat4(a.v[0]*d, a.v[1]*d, a.v[2]*d, a.v[3]*d);
 }
 
-mat4 GLUI::operator*(float d, const mat4 &a)
+mat4 GLUI::operator*(double d, const mat4 &a)
 {
     return a*d;
 }
 
-mat4 GLUI::operator/(const mat4 &a, float d)
+mat4 GLUI::operator/(const mat4 &a, double d)
 {
     return mat4(a.v[0]/d, a.v[1]/d, a.v[2]/d, a.v[3]/d);
 }
@@ -1511,11 +1511,11 @@ mat3 translation2D(const vec2 &v)
         vec3(0.0, 0.0, 1.0));
 }
 
-mat3 rotation2D(const vec2 &Center, float angleDeg)
+mat3 rotation2D(const vec2 &Center, double angleDeg)
 {
-    float angleRad = (float) (angleDeg * M_PI / 180.0);
-    float c = (float) cos(angleRad);
-    float s = (float) sin(angleRad);
+    double angleRad = (double) (angleDeg * M_PI / 180.0);
+    double c = (double) cos(angleRad);
+    double s = (double) sin(angleRad);
 
     return mat3(
         vec3(c,    -s, Center[VX] * (1.0f-c) + Center[VY] * s),
@@ -1549,12 +1549,12 @@ mat4 translation3D(const vec3 &v)
         vec4(0.0, 0.0, 0.0, 1.0));
 }
 
-mat4 rotation3D(const vec3 &Axis, float angleDeg)
+mat4 rotation3D(const vec3 &Axis, double angleDeg)
 {
-    float angleRad = (float) (angleDeg * M_PI / 180.0);
-    float c = (float) cos(angleRad);
-    float s = (float) sin(angleRad);
-    float t = 1.0f - c;
+    double angleRad = (double) (angleDeg * M_PI / 180.0);
+    double c = (double) cos(angleRad);
+    double s = (double) sin(angleRad);
+    double t = 1.0f - c;
 
     vec3 axis(Axis);
     axis.normalize();
@@ -1575,11 +1575,11 @@ mat4 rotation3D(const vec3 &Axis, float angleDeg)
         vec4(0.0, 0.0, 0.0, 1.0));
 }
 
-mat4 rotation3Drad(const vec3 &Axis, float angleRad)
+mat4 rotation3Drad(const vec3 &Axis, double angleRad)
 {
-    float c = (float) cos(angleRad);
-    float s = (float) sin(angleRad);
-    float t = 1.0f - c;
+    double c = (double) cos(angleRad);
+    double s = (double) sin(angleRad);
+    double t = 1.0f - c;
 
     vec3 axis(Axis);
     axis.normalize();
@@ -1609,7 +1609,7 @@ mat4 scaling3D(const vec3 &scaleVector)
         vec4(0.0, 0.0, 0.0, 1.0));
 }
 
-mat4 perspective3D(float d)
+mat4 perspective3D(double d)
 {
     return mat4(
         vec4(1.0f, 0.0f, 0.0f,   0.0f),
